requirements.txt
bcrypt==4.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
Django==2.2.28
django-registration-redux==2.2
idna==3.10
pillow==11.1.0
pytz==2024.2
requests==2.32.3
sqlparse==0.5.3
urllib3==2.3.0
whitenoise==6.9.0
